moved integrated algorithms
